Add all passing tests
