+++ app/controllers/admin/dashboard_controller.rb
@@ -1,11 +1,5 @@
-class Admin::DashboardController < ApplicationController
-  before_action :require_admin
-  
+class Admin::DashboardController < Admin::BaseController
+
   def index
   end
-
-  private
-    def require_admin
-      render file: "/public/404" unless current_admin?
-    end
 end
