+++ app/controllers/merchants/base_controller.rb
@@ -1,0 +1,8 @@
+class Merchants::BaseController < ApplicationController
+  before_action :require_merchant
+
+
+  def require_merchant
+    render file: "/public/404" unless current_merchant?
+  end
+end
