+++ app/controllers/users_controller.rb
@@ -5,21 +5,10 @@ class UsersController < ApplicationController
   end
 
   def create
-    require "pry"; binding.pry
     @user = User.new(user_params)
     if @user.save
       flash[:success] = "Welcome, #{@user.name}"
       redirect_to '/profile'
-      # if @user.role == 2
-      #   flash[:success] = "Welcome Admin #{@user.name}!"
-      #   redirect_to '/admin/dashboard'
-      # elsif @user.role == 1
-      #   flash[:success] = "Welcome Merchant, #{@user.name}!"
-      #   redirect_to '/merchants/dashboard'
-      # else
-      #   flash[:success] = "Welcome, #{@user.name}"
-      #   redirect_to '/profile'
-      # end
     elsif User.email_exists?(@user.email)
       flash[:failure] = @user.errors.full_messages.to_sentence
       @user.email = nil
